moved `tests` folder into `src` folder
